// sdk/nodejs/ec2/enums.ts
export const ProtocolType = {
    All: "all",
    TCP: "tcp",
    UDP: "udp",
    ICMP: "icmp",
} as const;

export type ProtocolType = (typeof ProtocolType)[keyof typeof ProtocolType];